+++ lcd_engine.py
@@ -1,11 +1,15 @@
 from PIL import Image
 from RPLCD.i2c import CharLCD
+from gpiozero import TonalBuzzer
+from gpiozero.tones import Tone
+from gpiozero.exc import BadPinFactory
 import time
 from gpiozero import Button
 from gpiozero import Device
 from gpiozero.pins.pigpio import PiGPIOFactory
 
 Device.pin_factory = PiGPIOFactory()
+
 
 lcd = CharLCD(i2c_expander="PCF8574", address=0x27, port=1, cols=16, rows=2, dotsize=8)
 a_button = Button(6)
@@ -29,7 +33,44 @@ class Engine:
         byte_map = [int("".join(map(str, row)), 2) for row in matrix]
 
         lcd.create_char(number, byte_map)
+    class Sound:
+        def __init__(self, music = 'default', soundEffects: list[str] = []):
+            try: 
+                self.buzzer: TonalBuzzer = TonalBuzzer(26)
+            except BadPinFactory:
+                self.buzzer = None
+                print("buzzer setup failed")
+            self.currentNoteIndex = 0
+            self.soundEffects = {}
 
+            for effectName in soundEffects: 
+                with open(f"assets/soundeffects/{effectName}.txt") as f:
+                    notes = f.read().strip().split()
+                    self.soundEffects[effectName] = { q : notes[q] for q in range(len(notes))}
+                    f.close()
+
+            with open(f"assets/music/{music}.txt") as f:
+                notes = f.read().strip().split()
+                self.soundtrackLength = len(notes)
+                self.musicNotes = {i: notes[i] for i in range(self.soundtrackLength)}
+                f.close()
+
+        def playSoundEffect(self, effectName: str):
+            if not self.buzzer:
+                return
+               
+            effectNotes: list[str] = self.soundEffects[effectName]
+            for i in range(len(effectNotes)):
+                # TODO: determine transition step value to reduce choppiness
+                self.buzzer.play(Tone(effectNotes[i]))   
+
+        # play the current note of the soundtrack. cycle to beginning when finished.
+        def playNote(self): 
+            if not self.buzzer:
+                return
+            # TODO: determine transition step value to reduce choppiness
+            self.buzzer.play(Tone(self.musicNotes[self.currentNoteIndex]))
+            self.currentNoteIndex = (self.currentNoteIndex + 1 ) % self.soundtrackLength
     class GameObject:
         x = 0
         y = 0
